streamlit_base/requirements.txt
streamlit==1.40.0
streamlit-drawable-canvas==0.9.3
streamlit-option-menu==0.3.13
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
pathlib
typing-extensions